test size of pkl'd trie w loaded words vs .txt
one bottle neck for the system is how long it takes to
load in words (esp on smaller machines). it can take
_minutes_ to read in the princetown wordnet which
has ~70000 entries. this may (probably will) be a problem
when reading in multiple dictionaries and creating multiple
tries for a forest.

so, we can test tradeoff between size of file and read in speed.

serialized tries after dictionaries have been read in (at least
for the princetown word net) takes 100x more bytes. but reading
in a pkl'd file instead of line by line is ~100x faster.